src/backend/requirements.txt
Flask==2.3.2
Flask-MySQL==1.5.1
Werkzeug>=2.3.3
mysql-connector-python==8.0.33
blinker>=1.6.2